Updated Scripts folder 3/29/2024.  Edited PlayerMovement and Platform_Movement
Edited the overall speed of first game(Dropping).  After asking others to playtest the game, general results showed that the game was too slow to increment speed and game dragged on too long.  Fixed by increasing initial movement of platforms and amount that it increments by.  Balanced by increasing user movement by a small amount to allow the player to live for a reasonable amount of time